+++ examples/stdin.rs
@@ -1,32 +1,29 @@
 extern crate futures;
-extern crate tokio_core;
-extern crate tokio_io;
+extern crate tokio;
 extern crate tokio_file_unix;
 
 use futures::{Future, future};
 
 fn main() {
-    let mut core = tokio_core::reactor::Core::new().unwrap();
-    let handle = core.handle();
-
     // convert stdin into a nonblocking file;
     // this is the only part that makes use of tokio_file_unix
-    let stdin = std::io::stdin();
-    let file = tokio_file_unix::StdFile(stdin.lock());
+    let file = tokio_file_unix::raw_stdin().unwrap();
     let file = tokio_file_unix::File::new_nb(file).unwrap();
-    let file = file.into_reader(&handle).unwrap();
+    let file = file.into_reader(&tokio::reactor::Handle::current()).unwrap();
 
     println!("Type something and hit enter!");
-    core.run(future::loop_fn((file, Vec::new()), |(file, line)| {
+    tokio::run(future::loop_fn((file, Vec::new()), |(file, line)| {
         // read each line
-        tokio_io::io::read_until(file, b'\n', line).map(|(file, mut line)| {
+        tokio::io::read_until(file, b'\n', line).map(|(file, mut line)| {
 
             // demonstrate that the event loop isn't blocked by I/O!
-            let one_sec = std::time::Duration::new(1, 0);
-            handle.spawn(
-                tokio_core::reactor::Timeout::new(one_sec, &handle).unwrap()
+            let one_sec_from_now =
+                std::time::Instant::now()
+                + std::time::Duration::new(1, 0);
+            tokio::spawn(
+                tokio::timer::Delay::new(one_sec_from_now)
                 .map_err(|_| ())
-                .map(|()| println!(" ... timeout works!"))
+                .map(|()| eprintln!(" ... timeout works!"))
             );
 
             if line.ends_with(b"\n") {
@@ -40,5 +37,5 @@ fn main() {
                 future::Loop::Break(())
             }
         })
-    })).unwrap();
+    }).map_err(|e| panic!("{:?}", e)));
 }